add armv7a and x86 build

+++ src/core/base/7zArchive.cpp
@@ -10,7 +10,7 @@ extern "C" {
 #include "StorageImpl.h"
 
 static ISzAlloc allocImp = {
-	[](void* p, unsigned long size) -> void *{ return malloc(size); },
+	[](void* p, size_t size) -> void *{ return malloc(size); },
 	[](void* p, void *addr) { free(addr); }
 };
